Work around a portability issue for now
We'll maybe need a new compiler flag or something. I don't like leaving
something like this enabled by default.

+++ src/Back.cpp
@@ -46,8 +46,10 @@ BOOL InitBack(const char *fName, int type)
 	// This is ridiculously platform-dependant:
 	// It should break on big-endian CPUs, and platforms
 	// where short isn't 16-bit and long isn't 32-bit.
-	short bmp_header_buffer[7];
-	long bmp_header_buffer2[10];
+//	short bmp_header_buffer[7];
+//	long bmp_header_buffer2[10];
+	int16_t bmp_header_buffer[7];
+	int32_t bmp_header_buffer2[10];	// We'll need a better solution when we stop using stdint.h
 
 	fread(bmp_header_buffer, 14, 1, fp);
